refactor: re-structure and make clippy happy
