+++ spog/ui/src/console/mod.rs
@@ -1,3 +1,5 @@
+mod auth;
+
 use crate::{
     about,
     backend::Endpoint,
@@ -72,7 +74,7 @@ pub fn console() -> Html {
     );
 
     let auth = use_auth_state();
-    let auth = use_memo(|auth| from_auth(auth), auth);
+    let auth = use_memo(auth::from_auth, auth);
 
     let agent = use_auth_agent().expect("Requires OAuth2Context component in parent hierarchy");
     let onlogout = use_callback(
@@ -139,55 +141,7 @@ pub fn console() -> Html {
     )
 }
 
-struct FromAuth {
-    avatar: Html,
-    account_url: Option<String>,
-    name: String,
-    username: String,
-}
-
-fn from_auth(auth: &Option<OAuth2Context>) -> FromAuth {
-    let (_email, account_url, username, name) = match auth.as_ref().and_then(|auth| auth.claims()) {
-        Some(claims) => {
-            let account_url = {
-                let mut issuer = claims.issuer().url().clone();
-                if let Ok(mut paths) = issuer
-                    .path_segments_mut()
-                    .map_err(|_| anyhow::anyhow!("Failed to modify path"))
-                {
-                    paths.push("account");
-                }
-                issuer.to_string()
-            };
-
-            let username = claims
-                .preferred_username()
-                .map(|s| s.as_ref())
-                .unwrap_or_else(|| claims.subject().as_str())
-                .to_string();
-
-            let name = claims
-                .name()
-                .and_then(|name| name.get(None))
-                .map(|s| s.to_string())
-                .unwrap_or_else(|| username.clone());
-
-            (claims.email(), Some(account_url), username, name)
-        }
-        None => (None, None, String::default(), String::default()),
-    };
-
-    // TODO: for now use the default, consider using the profile image
-    let src = "assets/images/img_avatar.svg".to_string();
-
-    FromAuth {
-        avatar: html!(<Avatar {src} alt="avatar" size={AvatarSize::Small} />),
-        account_url,
-        name,
-        username,
-    }
-}
-
+/// Drop down switch entry for dark mode
 #[function_component(DarkModeEntry)]
 fn dark_mode_entry() -> Html {
     html!(
